Check for file / link existence before removal

+++ linux-bootstrap.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
-rm -rf ~/.vim
-rm ~/.vimrc
+if [ -e ~/.vim ] || [ -L ~/.vim ]
+then
+	rm -rf ~/.vim
+fi
+if [ -e ~/.vimrc ] || [ -L ~/.vimrc ]
+then
+	rm -f ~/.vimrc
+fi
 ln -s ${PWD##*/}/.vim ~/.vim
 ln -s ${PWD##*/}/.vimrc ~/.vimrc
 git submodule update --init
